renamed flat-network-controller to arktos-network-controller
