+++ cmd/arktos-network-controller/BUILD
@@ -2,11 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["flat-network-controller.go"],
-    importpath = "k8s.io/kubernetes/cmd/flat-network-controller",
+    srcs = ["network-controller.go"],
+    importpath = "k8s.io/kubernetes/cmd/arktos-network-controller",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/flat-network-controller/app:go_default_library",
+        "//cmd/arktos-network-controller/app:go_default_library",
         "//staging/src/k8s.io/arktos-ext/pkg/generated/clientset/versioned:go_default_library",
         "//staging/src/k8s.io/arktos-ext/pkg/generated/informers/externalversions:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
@@ -17,7 +17,7 @@ go_library(
 )
 
 go_binary(
-    name = "flat-network-controller",
+    name = "arktos-network-controller",
     embed = [":go_default_library"],
     visibility = ["//visibility:public"],
 )
@@ -33,7 +33,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//cmd/flat-network-controller/app:all-srcs",
+        "//cmd/arktos-network-controller/app:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
